requirements.txt
fastapi==0.95.2
python-multipart==0.0.6
uvicorn==0.22.0
rasterio==1.3.7
aiofiles==23.1.0
Pillow==9.5.0
numpy==1.24.3